pyproject.toml
[tool.poetry]
name = "demo2"
version = "0.1.0"
description = ""
authors = ["Chathuranga Bandara <chathurangapayment@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
farm-haystack = {extras = ["inference"], version = "^1.24.1"}
setuptools = "^69.1.1"
selenium = "^4.18.1"
torch = "^2.2.1"
nltk = "^3.8.1"
sqlalchemy = "^2.0.27"
streamlit = "^1.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
